patchkernel: define patch dimension as a short integer

+++ src/levelset/levelSetCartesian.cpp
@@ -67,11 +67,11 @@ double LevelSetCartesian::computeCellIncircle(long id) {
 
     BITPIT_UNUSED(id);
 
-    int dim = m_cartesian->getDimension();
+    short dim = m_cartesian->getDimension();
     std::array<double,3> spacing = m_cartesian->getSpacing();
     double minSpacing = std::numeric_limits<double>::max() ;
 
-    for(int i=0; i<dim; ++i){
+    for(short i=0; i<dim; ++i){
         minSpacing = std::min( minSpacing, spacing[i] );
     }
 
@@ -87,11 +87,11 @@ double LevelSetCartesian::computeCellCircumcircle( long id ) {
 
     BITPIT_UNUSED(id);
 
-    int dim = m_cartesian->getDimension();
+    short dim = m_cartesian->getDimension();
     std::array<double,3> spacing = m_cartesian->getSpacing();
     double diagonalSquare = 0.0;
 
-    for(int i=0; i<dim; ++i){
+    for(short i=0; i<dim; ++i){
         diagonalSquare += spacing[i]*spacing[i];
     }
 
@@ -112,7 +112,7 @@ bool LevelSetCartesian::intersectCellPlane( long id, const std::array<double,3> 
     std::array<double,3> minPoint( centroid -0.5*spacing );
     std::array<double,3> maxPoint( centroid +0.5*spacing );
 
-    int dim = m_cartesian->getDimension();
+    short dim = m_cartesian->getDimension();
     return CGElem::intersectPlaneBox( root, normal, minPoint, maxPoint, dim);
 }
