+++ src/surfunstructured/surfunstructured.hpp
@@ -40,13 +40,13 @@ public:
 
     // Constructors
     SurfUnstructured();
-    SurfUnstructured(int patch_dim, int space_dim);
-    SurfUnstructured(int id, int patch_dim, int space_dim);
+    SurfUnstructured(short patch_dim, short space_dim);
+    SurfUnstructured(int id, short patch_dim, short space_dim);
     SurfUnstructured(std::istream &stream);
 #if BITPIT_ENABLE_MPI==1
     SurfUnstructured(MPI_Comm communicator);
-    SurfUnstructured(int patch_dim, int space_dim, MPI_Comm communicator);
-    SurfUnstructured(int id, int patch_dim, int space_dim, MPI_Comm communicator);
+    SurfUnstructured(short patch_dim, short space_dim, MPI_Comm communicator);
+    SurfUnstructured(int id, short patch_dim, short space_dim, MPI_Comm communicator);
     SurfUnstructured(std::istream &stream, MPI_Comm communicator);
 #endif
